% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestry_measures.R
\name{ba}
\alias{ba}
\title{Calculate basal area}
\usage{
ba(dbh, na.rm = TRUE)
}
\arguments{
\item{dbh}{numeric - vector representing measure of diameter at breast height}

\item{na.rm}{boolean - should NAs be ignored.}
}
\description{
This function calculates basal area from DBH using simple circular assumption.
Out put is in squared units of inputs.
}
\examples{
library(landecoutils)
print(ba(c(.101, .064, NA, 0.023), na.rm=TRUE))
}
