% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_lidR_ctgs.R
\name{check_for_duplicate_scans}
\alias{check_for_duplicate_scans}
\title{Check LAS Catalog for duplicate scans, this will reduce errors elsewhere.}
\usage{
check_for_duplicate_scans(ctg)
}
\arguments{
\item{ctg}{a LAS Catalog}
}
\description{
This function takes a LAS Catalog and throws an error
if any duplicate scan positions are detected. These files are sometimes
generated if a bad scan was taken in the field and a second scan
was generated from the same position
}
