% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_lidR_ctgs.R
\name{clip_las_catalog}
\alias{clip_las_catalog}
\title{Clip a LAS catalog to a boundary to speed up processing}
\usage{
clip_las_catalog(dir, bnd, output_dir)
}
\arguments{
\item{dir}{a directory of LAS/LAZ files}

\item{bnd}{an \code{sf} object denoting the region of interest. Only UTM units tested}

\item{output_dir}{directory to output LAS tiles}
}
\description{
This function takes an input directory of TLS scans that are potentially
overlapping and a region of interest (\code{bnd}). The function loads and clips
all las data to \code{bnd} and outputs LAS tiles with an extent
matching \code{bnd}.
}
