% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_lidR_ctgs.R
\name{correct_Reflectance_Xu2017}
\alias{correct_Reflectance_Xu2017}
\title{Correct Return Reflectance based on Distance using Xu et al. 2017}
\usage{
correct_Reflectance_Xu2017(las)
}
\arguments{
\item{las}{a \code{LAS} object containing attributes for \code{Amplitude} and \code{Distance}}
}
\description{
This function takes a \code{LAS} input containing the fields \code{Amplitude} in dB and
\code{Distance} in meteres. It applies a distance based correction using the methods
of Xu et al. 2017. The formula only applies the correction based on distance, but
not incident angle. The study used a REIGL VZ400i, but less sure if this applies
to other scanners.
}
\examples{
## NOT RUN ##
#las = readLAS('las1.laz')
# corr = correct_Reflectance_Xu2017(las)
# par(mfrow = c(1,2))
# plot(las$Amplitude ~ las$Distance)
# plot(corr, las$Distance)
}
