% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_lidR_ctgs.R
\name{find_las_centroid}
\alias{find_las_centroid}
\title{Find centroid of large *.las}
\usage{
find_las_centroid(las, res = 1)
}
\arguments{
\item{las}{a \code{lidR::LAS} object for which you want to find a centroid}

\item{res}{resolution to thin the point cloud for efficiency}
}
\description{
This function takes a \code{lidR::LAS} object, and reads it in quickly by dropping
most points using \code{subsample}. The function returns an \code{sf::POINT} object
containing the coordinates of the centroid. \emph{Note} that centroid is only
approximate and may be imprecise if subsample is large (i.e., > 1000) or
LAS point density is low.
}
\examples{
# Load large LAS file and identify the centroid (i.e., scan location)
library(lidR)
library(sf)
#' las = readLAS('large_las_file.las')
cent = find_las_centroid(las)
plot(cent$geom)
}
