% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_lidR_ctgs.R
\name{get_ctg_crs}
\alias{get_ctg_crs}
\title{Get Coordinate Reference System (CRS) from LASCatalog}
\usage{
get_ctg_crs(ctg)
}
\arguments{
\item{ctg}{a \code{LAScatalog} object containing LAS files}
}
\description{
This function takes a LASCatalog as input and returns a \code{crs} object
representing the coordinate reference system of the LASCatalog
}
\examples{
# Load LAScatalog and capture the crs
ctg = readLAScatalog('path/to/LASfiles/')
myCRS = get_crg_crs(ctg)
print(myCRS)
}
