% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_lidR_ctgs.R
\name{get_dem_csm_chm}
\alias{get_dem_csm_chm}
\title{Create elevation products from a LAS tile}
\usage{
get_dem_csm_chm(las, res = 0.5)
}
\arguments{
\item{las}{LAS object from lidR package to create model}
}
\description{
This function takes a LAS object and returns a digital elevation model
(dem), a canopy surface model (csm), and a canopy height model (chm)
using default algorithms in lidR. Can be used with lidR::catalog_map.
Returns a 3-layered SpatRast object with names dem, csm, chm.
}
\examples{
library(terra)
las = readLAS('E:/mylas.laz')
elev = get_cdem_csm_chm(las)
plot(elev$dem)
plot(elev$csm)
plot(elev$chm)

ctg = readLASCatalog('E:/my/las/dir/')
elev = catalog_map(ctg, get_dem_csm_chm)
plot(elev$dem)
plot(elev$csm)
plot(elev$chm)

}
