% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_lidR_ctgs.R
\name{index_las}
\alias{index_las}
\title{Check for and create a lax index from a directory of LAS files}
\usage{
index_las(las_dir, n_cores)
}
\arguments{
\item{las_dir}{path to a directory containing .LAS or .LAZ files to index}
}
\description{
This function takes a directory of LAS files and checks to see if they are
accomponied by .LAX index files. Indexing greatly increases the processing
speed for many tasks.
}
\examples{
index_las('E:/my/las/dir/')
}
