% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_lidR_ctgs.R
\name{las_add_scanner_distance}
\alias{las_add_scanner_distance}
\title{Add attribute \code{Distance} to TLS returns based on distance from scanner}
\usage{
las_add_scanner_distance(las_filename, scanner_ht = 1.75, subsample = 10000)
}
\arguments{
\item{las_filename}{path to a \code{.las} or \code{.laz} file}

\item{scanner_ht}{height of scanner above ground. Defaults to 1.75}

\item{subsample}{subsamplign factor used in \code{find_las_centroid()} to reduce lidar resoultion. Defaults to 1e05.}
}
\description{
This function takes the path of a .las or .laz file and return a \code{LAS} object
containing the new attribute \code{Distance} represetning the distance to the scanner
The scanner location is determined by (1) using the \code{find_las_centroid()} function,
then (2) locates the scanner elevation by adding the distance from the \code{scanner_ht}
parameter to the elevation of the centroid. Elevation is determined using the
\code{lidR::classify_ground} and \code{lidR::rasterize_terrain} functions.
}
\examples{

las_fn = 'data.laz'
las = las_add_scanner_distance(las_fn)
lidR::writeLAS(las, 'data_withDistance.laz')
}
