% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_pts.R
\name{make_scan_points}
\alias{make_scan_points}
\title{Create points for scan locations around a central point}
\usage{
make_scan_points(input_shp, output_shp, dist, dirs = c(0, 90, 180, 270))
}
\arguments{
\item{input_shp}{path to a .SHP containing points}

\item{output_shp}{path to a .SHP to output points}

\item{dist}{distance (in meters) of desired distance to each point}

\item{dirs}{directions of desired new points (in degrees)}
}
\description{
This function takes a shapefile containing plot locations,
and creates new points in the specified directions at a constant
distance. It returns the original points (newid=0) plus the
surrounding pionts (newid > 0)
}
\examples{
in_file = 'R:/landscape_ecology/test.shp'
out_file = 'R:/landscape_ecology/output.shp'
dirs = c(0,90,180,270)
dist = 20
make_scan_points(in_file, out_file, dist, dirs)

}
