% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanner_segment.R
\name{segment_with_spanner}
\alias{segment_with_spanner}
\title{Create a stem map from a TLS scan tile using spanner.}
\usage{
segment_with_spanner(las, bnd, stemmap_shp, tree_seg_las = NULL, threads = 0.5)
}
\arguments{
\item{las}{a \code{LAS} object not much larger than 0.2 ha}

\item{bnd}{an \code{sf} object representing the \emph{interior} boundary of the study area.
Generally \code{bnd} is smaller than the \code{bbox} of \code{las} to adjust edge effects}

\item{stemmap_shp}{a path name (.shp) with the desired output of the resulting stem map.}

\item{tree_seg_las}{a path name (.las) for the output of segmented trees}

\item{threads}{see \code{lidR::set_lidR_threads}. Positive scalar. Default 0 means use all
CPU available. Values > 1 mean using n cores, values in (0, 1) mean using a
fraction of the cores e.g. 0.5 = half.}
}
\description{
This is code to make it more convenient to pre-process, stem map, and
segment TLS tiles. The function takes a merged TLS scan and a boundary
and generates a stem map (.shp) attributed with stem radius, and height.
You may optionally output a LAS with the segmented trees by specifying
a path name (*.las) in the argument \code{tree_seg_las}. If left NULL, only the
stem map will be written to disk. Note that there are large memory requirements
and you may encounter memory issues if you use this on areas
larger than 0.2 ha with high scan densities.
}
