% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_lidR_ctgs.R
\name{stitch_TLS_dir_to_LAS}
\alias{stitch_TLS_dir_to_LAS}
\title{Stitch TLS scans into single *.las}
\usage{
stitch_TLS_dir_to_LAS(
  ctg,
  out_las,
  roi,
  buffer = 10,
  max_scan_distance = 60,
  index = TRUE,
  scan_locations = NULL
)
}
\arguments{
\item{ctg}{a \code{LAScatalog} object containing LAS files to be stitched}

\item{out_las}{path for output *.las file}

\item{roi}{an \code{sf} object denoting the region of interest. UTM units preferred}

\item{buffer}{width of buffer to apply to \code{roi} using \code{sf::st_buffer}}

\item{max_scan_distance}{maximum distance from scan to be included. Evaluated using \code{find_las_centroid()}}

\item{index}{boolean. Also write a lax file to index the points in the files. see \code{lidR::writeLAS}}

\item{scan_locations}{(optional) an \code{sf} object of scan centroids index the same as \code{ctg}.
If \code{scan_locations} is \code{NULL} then scan locations will be found automatically}
}
\description{
This function takes an input directory of TLS scans that are overlapping
and a region of interest. The function loads, stitches, and clips
all las data to the region of interest and outputs into a single file.
Note that this should only be used on relatively small areas to avoid
very large file sizes and memory errors. Plot sizes < 0.2 ha are recommended
including buffer unless points are sparse. \emph{Note}: script will
keep only data columns which are common among all scans.
}
\examples{
# Load LAScatalog and clip to a region of interest specified by an sf object
ctg = lidR::readLAScatalog('path/to/LASfiles/')
roi = sf::st_read('plot_boundary.shp')
stitch_TLS_dir_to_LAS(ctg, 'output_path.las', roi)
}
