% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_lidR_ctgs.R
\name{stitch_TLS_dir_to_LAS_tiles}
\alias{stitch_TLS_dir_to_LAS_tiles}
\title{Combine overlapping TLS scans into tiled LAS scene}
\usage{
stitch_TLS_dir_to_LAS_tiles(
  ctg,
  out_dir,
  tile_size = 30,
  n_cores = 1,
  buffer = 10,
  max_scan_distance = 60,
  bnd = NULL,
  index = TRUE,
  scan_locations = NULL,
  timeout = 600
)
}
\arguments{
\item{ctg}{a \code{LAScatalog} object containing overlapping TLS scans}

\item{out_dir}{directory to output LAS tiles (no ending '/' in path).}

\item{n_cores}{number of cores to use in parallel processing}

\item{buffer}{width of buffer to apply to \code{bnd} using \code{sf::st_buffer} This
allows inclusion of TLS data outside study area to address edge effects.}

\item{max_scan_distance}{maximum distance from scan to be included. Evaluated using \code{find_las_centroid()}}

\item{bnd}{an \code{sf} object denoting the region of interest. Only UTM units tested}

\item{index}{boolean. Also write a lax file to index the points in the files. see \code{lidR::writeLAS}}

\item{timeout}{numeric in seconds. Number of seconds for timeout on a tile
before proceeding to the next one. This avoids cores getting hung up on
problem tiles, so that that can be trouble shooted separately. This may not
be necessary in later versions.}
}
\description{
This function takes an input directory of TLS scans that are overlapping
and a region of interest (e.g., large plot). The function loads, stitches, and clips
all las data to the region of interest and outputs LAS tiles of a user specified
size. \emph{Note}: when binding overlapping scan sections, script will
keep only data columns which are common among all scans.
}
\examples{
# Load LAScatalog, clip, and tile for a large area specified by an sf object
ctg = lidR::readLAScatalog('path/to/LASfiles/')
bnd = sf::st_read('plot_boundary.shp')
stitch_TLS_dir_to_LAS_tiles(ctg, 'output_tiles', bnd, tile_size = 30)
}
