% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_lidR_ctgs.R
\name{tile_large_LAS}
\alias{tile_large_LAS}
\title{Function to clip large LAS object into small pieces}
\usage{
tile_large_LAS(
  las_path,
  output_directory,
  chunk_size = 100,
  chunk_alighment = c(0, 0)
)
}
\arguments{
\item{las_path}{character - path to large LAS file}

\item{output_directory}{character - path to directory for output files.
Outputs will be automatically named XLEFT_YBOTTOM.las'}

\item{chunk_size}{numeric size of the chunks to break the file into}

\item{chunk_alignment}{coordinates of origin of grid to which files}
}
\description{
This function clips a large LAS object such as that from a
UAV-derived point cloud into smaller chunks using the
\code{catalog_map} function.
}
\examples{
library(lidR)
library(landecoutils)
large_file = 'C:/user/me/large_file.las'
output_location = 'C:/user/me/outputs/'
tile_large_LAS(large_file, output_location)
}
