% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_3d_voxel.R
\name{clip_voi}
\alias{clip_voi}
\title{Clip a LAS object to a volume of interest}
\usage{
clip_voi(las, voi, res)
}
\arguments{
\item{las}{a \code{LAS} usually with a large extent and/or high point density}

\item{voi}{a \code{LAS} object often with lower resolution and extent from which
to generate a volume to clip \code{las}}

\item{`res`}{the resolution at which the \code{voi} is generated}
}
\description{
This code is a 3-dimensional version of \code{lidR::clip_roi} which interprets
a smaller las object as a volume of interest and clips a larger LAS object
to that volume. Some segmentation algorithms result in lower resolution
LAS objects than the original files, so this function can be used to clip
a low resolution segmentated LAS object and retrieve the volume specified from
a higher resolution LAS.
}
\examples{
library(landecoutils)
library(lidR)

# Large full resolution LAS object
full_las = readTLSLAS('path/to/highres.las')

# smaller lower resolution LAS object
voi = readTLSLAS('path/to/lowres.las')

# view point density an extent of each las
nrow(full_las@data);extent(full_las)
nrow(voi@data);extent(voi)

# Clip voi and view point density and extent
high_res_voi = clip_voi(full_las, voi, res=0.5)
nrow(high_res_voi@data); extent(high_res_voi)
}
