% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_lidR_ctgs.R
\name{compress_las}
\alias{compress_las}
\title{Compress a folder of LAS files to LAZ format.}
\usage{
compress_las(las_dir, n_cores, index = TRUE, delete_old = FALSE)
}
\arguments{
\item{las_dir}{path to a directory containing .LAS files to compress}

\item{n_cores}{number of cores to create doSNOW cluster}
}
\description{
This function takes a directory of LAS files and compresses them to LAZ.
The original LAS files are deleted assuming you want to save space.
}
\examples{
## NOT RUN ##
# compress_las('E:/my/las/dir/', n_cores=2)
}
