% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_lidR_ctgs.R
\name{find_ctg_centroids}
\alias{find_ctg_centroids}
\title{Find scan locations from a \code{LAScatalog}}
\usage{
find_ctg_centroids(ctg, n_cores = 1, subsample = 10000)
}
\arguments{
\item{ctg}{a \code{lidR::LAScatalog} object for which you want to find a centroids of all scan locations}

\item{n_cores}{a number of cores to use for parallel processing.}

\item{subsample}{drop every \code{nth} return to speed up processing time. See \code{las_find_centroids}}
}
\description{
This function does the same as \code{las_find_centroids} but it works on a
\code{LAScatalog} and returns an \code{sf} object of point lcoations of scan centroids.
Result of this function can be used as an input to \code{stitch_TLS_dir_to_LAS} and
\code{stitch_TLS_dir_to_LAS_tile} to increase speed.
}
\examples{
library(sf)
ctg = readTLScatalog('path/to/las/catalog')
scan_locations = find_ctg_centroids(ctg, n_cores=4)
plot(scan_locations$geom)
}
